refactor(trivial): depress warning

+++ test/benchmark.cpp
@@ -40,21 +40,25 @@ uint64_t latency_th_all[LATENCY_WINDOWS];
 Tree *tree;
 DSM *dsm;
 
-inline Key to_key(uint64_t k) {
+inline Key to_key(uint64_t k)
+{
   return (CityHash64((char *)&k, sizeof(k)) + 1) % kKeySpace;
 }
 
-class RequsetGenBench : public RequstGen {
+class RequsetGenBench : public RequstGen
+{
 
 public:
   RequsetGenBench(int coro_id, DSM *dsm, int id)
-      : coro_id(coro_id), dsm(dsm), id(id) {
+      : coro_id(coro_id), dsm(dsm), id(id)
+  {
     seed = rdtsc();
     mehcached_zipf_init(&state, kKeySpace, zipfan,
                         (rdtsc() & (0x0000ffffffffffffull)) ^ id);
   }
 
-  Request next() override {
+  Request next() override
+  {
     Request r;
     uint64_t dis = mehcached_zipf_next(&state);
 
@@ -76,14 +80,16 @@ private:
   struct zipf_gen_state state;
 };
 
-RequstGen *coro_func(int coro_id, DSM *dsm, int id) {
+RequstGen *coro_func(int coro_id, DSM *dsm, int id)
+{
   return new RequsetGenBench(coro_id, dsm, id);
 }
 
 Timer bench_timer;
 std::atomic<int64_t> warmup_cnt{0};
 std::atomic_bool ready{false};
-void thread_run(int id) {
+void thread_run(int id)
+{
 
   bindCore(id);
 
@@ -92,26 +98,30 @@ void thread_run(int id) {
   uint64_t all_thread = kThreadCount * dsm->getClusterSize();
   uint64_t my_id = kThreadCount * dsm->getMyNodeID() + id;
 
-  printf("I am %d\n", my_id);
+  printf("I am %llu\n", (unsigned long long)my_id);
 
-  if (id == 0) {
+  if (id == 0)
+  {
     bench_timer.begin();
   }
 
-  for (uint64_t i = 1; i < kKeySpace; ++i) {
-    if (i % all_thread == my_id) {
+  for (uint64_t i = 1; i < kKeySpace; ++i)
+  {
+    if (i % all_thread == my_id)
+    {
       tree->insert(to_key(i), i * 2);
     }
   }
 
   warmup_cnt.fetch_add(1);
 
-  if (id == 0) {
+  if (id == 0)
+  {
     while (warmup_cnt.load() != kThreadCount)
       ;
     printf("node %d finish\n", dsm->getMyNodeID());
     dsm->barrier("warm_finish");
-  
+
     uint64_t ns = bench_timer.end();
     printf("warmup time %lds\n", ns / 1000 / 1000 / 1000);
 
@@ -138,9 +148,11 @@ void thread_run(int id) {
                       (rdtsc() & (0x0000ffffffffffffull)) ^ id);
 
   Timer timer;
-  while (true) {
+  while (true)
+  {
 
-    if (need_stop) {
+    if (need_stop)
+    {
       while (true)
         ;
     }
@@ -160,14 +172,18 @@ void thread_run(int id) {
     Value v;
 
     timer.begin();
-    if (rand_r(&seed) % 100 < kReadRatio) { // GET
+    if (rand_r(&seed) % 100 < kReadRatio)
+    { // GET
       tree->search(key, v);
-    } else {
+    }
+    else
+    {
       v = 12;
       tree->insert(key, v);
     }
     auto us_10 = timer.end() / 100;
-    if (us_10 >= LATENCY_WINDOWS) {
+    if (us_10 >= LATENCY_WINDOWS)
+    {
       us_10 = LATENCY_WINDOWS - 1;
     }
     latency[id][us_10]++;
@@ -178,8 +194,10 @@ void thread_run(int id) {
 #endif
 }
 
-void parse_args(int argc, char *argv[]) {
-  if (argc != 4) {
+void parse_args(int argc, char *argv[])
+{
+  if (argc != 4)
+  {
     printf("Usage: ./benchmark kNodeCount kReadRatio kThreadCount\n");
     exit(-1);
   }
@@ -192,11 +210,14 @@ void parse_args(int argc, char *argv[]) {
          kReadRatio, kThreadCount);
 }
 
-void cal_latency() {
+void cal_latency()
+{
   uint64_t all_lat = 0;
-  for (int i = 0; i < LATENCY_WINDOWS; ++i) {
+  for (int i = 0; i < LATENCY_WINDOWS; ++i)
+  {
     latency_th_all[i] = 0;
-    for (int k = 0; k < MAX_APP_THREAD; ++k) {
+    for (int k = 0; k < MAX_APP_THREAD; ++k)
+    {
       latency_th_all[i] += latency[k][i];
     }
     all_lat += latency_th_all[i];
@@ -209,26 +230,32 @@ void cal_latency() {
   uint64_t th999 = all_lat * 999 / 1000;
 
   uint64_t cum = 0;
-  for (int i = 0; i < LATENCY_WINDOWS; ++i) {
+  for (int i = 0; i < LATENCY_WINDOWS; ++i)
+  {
     cum += latency_th_all[i];
 
-    if (cum >= th50) {
+    if (cum >= th50)
+    {
       printf("p50 %f\t", i / 10.0);
       th50 = -1;
     }
-    if (cum >= th90) {
+    if (cum >= th90)
+    {
       printf("p90 %f\t", i / 10.0);
       th90 = -1;
     }
-    if (cum >= th95) {
+    if (cum >= th95)
+    {
       printf("p95 %f\t", i / 10.0);
       th95 = -1;
     }
-    if (cum >= th99) {
+    if (cum >= th99)
+    {
       printf("p99 %f\t", i / 10.0);
       th99 = -1;
     }
-    if (cum >= th999) {
+    if (cum >= th999)
+    {
       printf("p999 %f\n", i / 10.0);
       th999 = -1;
       return;
@@ -236,7 +263,8 @@ void cal_latency() {
   }
 }
 
-int main(int argc, char *argv[]) {
+int main(int argc, char *argv[])
+{
 
   parse_args(argc, argv);
 
@@ -247,15 +275,18 @@ int main(int argc, char *argv[]) {
   dsm->registerThread();
   tree = new Tree(dsm);
 
-  if (dsm->getMyNodeID() == 0) {
-    for (uint64_t i = 1; i < 102400; ++i) {
+  if (dsm->getMyNodeID() == 0)
+  {
+    for (uint64_t i = 1; i < 102400; ++i)
+    {
       tree->insert(to_key(i), i * 2);
     }
   }
 
   dsm->barrier("benchmark");
 
-  for (int i = 0; i < kThreadCount; i++) {
+  for (int i = 0; i < kThreadCount; i++)
+  {
     th[i] = std::thread(thread_run, i);
   }
 
@@ -265,14 +296,16 @@ int main(int argc, char *argv[]) {
   timespec s, e;
   uint64_t pre_tp = 0;
   uint64_t pre_ths[MAX_APP_THREAD];
-  for (int i = 0; i < MAX_APP_THREAD; ++i) {
+  for (int i = 0; i < MAX_APP_THREAD; ++i)
+  {
     pre_ths[i] = 0;
   }
 
   int count = 0;
 
   clock_gettime(CLOCK_REALTIME, &s);
-  while (true) {
+  while (true)
+  {
 
     sleep(2);
     clock_gettime(CLOCK_REALTIME, &e);
@@ -280,13 +313,15 @@ int main(int argc, char *argv[]) {
                        (double)(e.tv_nsec - s.tv_nsec) / 1000;
 
     uint64_t all_tp = 0;
-    for (int i = 0; i < kThreadCount; ++i) {
+    for (int i = 0; i < kThreadCount; ++i)
+    {
       all_tp += tp[i][0];
     }
     uint64_t cap = all_tp - pre_tp;
     pre_tp = all_tp;
 
-    for (int i = 0; i < kThreadCount; ++i) {
+    for (int i = 0; i < kThreadCount; ++i)
+    {
       auto val = tp[i][0];
       // printf("thread %d %ld\n", i, val - pre_ths[i]);
       pre_ths[i] = val;
@@ -294,13 +329,15 @@ int main(int argc, char *argv[]) {
 
     uint64_t all = 0;
     uint64_t hit = 0;
-    for (int i = 0; i < MAX_APP_THREAD; ++i) {
+    for (int i = 0; i < MAX_APP_THREAD; ++i)
+    {
       all += (cache_hit[i][0] + cache_miss[i][0]);
       hit += cache_hit[i][0];
     }
 
     uint64_t fail_locks_cnt = 0;
-    for (int i = 0; i < MAX_APP_THREAD; ++i) {
+    for (int i = 0; i < MAX_APP_THREAD; ++i)
+    {
       fail_locks_cnt += lock_fail[i][0];
       lock_fail[i][0] = 0;
     }
@@ -311,34 +348,40 @@ int main(int argc, char *argv[]) {
     //  pattern
     uint64_t pp[8];
     memset(pp, 0, sizeof(pp));
-    for (int i = 0; i < 8; ++i) {
-      for (int t = 0; t < MAX_APP_THREAD; ++t) {
+    for (int i = 0; i < 8; ++i)
+    {
+      for (int t = 0; t < MAX_APP_THREAD; ++t)
+      {
         pp[i] += pattern[t][i];
         pattern[t][i] = 0;
       }
     }
 
     uint64_t hot_count = 0;
-    for (int i = 0; i < MAX_APP_THREAD; ++i) {
+    for (int i = 0; i < MAX_APP_THREAD; ++i)
+    {
       hot_count += hot_filter_count[i][0];
       hot_filter_count[i][0] = 0;
     }
 
     uint64_t hier_count = 0;
-    for (int i = 0; i < MAX_APP_THREAD; ++i) {
+    for (int i = 0; i < MAX_APP_THREAD; ++i)
+    {
       hier_count += hierarchy_lock[i][0];
       hierarchy_lock[i][0] = 0;
     }
 
     uint64_t ho_count = 0;
-    for (int i = 0; i < MAX_APP_THREAD; ++i) {
+    for (int i = 0; i < MAX_APP_THREAD; ++i)
+    {
       ho_count += handover_count[i][0];
       handover_count[i][0] = 0;
     }
 
     clock_gettime(CLOCK_REALTIME, &s);
 
-    if (++count % 3 == 0 && dsm->getMyNodeID() == 0) {
+    if (++count % 3 == 0 && dsm->getMyNodeID() == 0)
+    {
       cal_latency();
     }
 
@@ -347,12 +390,14 @@ int main(int argc, char *argv[]) {
 
     printf("%d, throughput %.4f\n", dsm->getMyNodeID(), per_node_tp);
 
-    if (dsm->getMyNodeID() == 0) {
+    if (dsm->getMyNodeID() == 0)
+    {
       printf("cluster throughput %.3f\n", cluster_tp / 1000.0);
 
       printf("cache hit rate: %lf\n", hit * 1.0 / all);
       printf("ACCESS PATTERN");
-      for (int i = 0; i < 8; ++i) {
+      for (int i = 0; i < 8; ++i)
+      {
         printf("\t%ld", pp[i]);
       }
       printf("\n");
